api_cdn_config/requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.1.2
django-extensions==2.1.3
django-filter==2.1.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
httpie==1.0.2
idna==2.8
isort==4.3.20
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
# mysqlclient==1.4.2.post1
openapi-codec==1.3.2
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
python-decouple==3.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
validators==0.13.0

wrapt==1.11.1
gunicorn==19.9.0
pymysql==0.9.3

drf_yasg
djangorestframework-link-header-pagination
packaging